third_party/blink/web_tests/wpt_internal/webgpu/canvas_webgpu_transfer/transferBackFromGPUTexture-first-throws.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-promise-test.js"></script>

<h1>Canvas2D WebGPU Transfer: transferToGPUTexture-first-throws</h1>
<p>Calling transferBackFromGPUTexture() without any preceding call to
transferToGPUTexture() should raise an exception.</p>

<script>

runCanvasTestsInWorker();

/**
 * Calling transferBackFromGPUTexture() without any preceding call to
 * transferToGPUTexture should raise an exception.
 */
canvasPromiseTest(async (canvas) => {
  const ctx = canvas.getContext('2d');

  try {
    ctx.transferBackFromGPUTexture();
    assert_unreached('transferBackFromGPUTexture should have thrown.');
  } catch (ex) {
    assert_true(ex instanceof DOMException);
    assert_equals(ex.name, 'InvalidStateError');
  }
}, 'Calling transferBackFromGPUTexture() without any preceding call to ' +
   'transferToGPUTexture() should raise an exception.');

</script>
